+++ app-backend/db/src/main/scala/meta/EnumMeta.scala
@@ -71,4 +71,18 @@ trait EnumMeta {
 
   implicit val taskStatusMeta: Meta[TaskStatus] =
     pgEnumString("task_status", TaskStatus.fromString, _.repr)
+
+  implicit val annotationProjectTypeMeta: Meta[AnnotationProjectType] =
+    pgEnumString(
+      "annotation_project_type",
+      AnnotationProjectType.fromString,
+      _.repr
+    )
+
+  implicit val tileLayerTypeMeta: Meta[TileLayerType] =
+    pgEnumString(
+      "tile_layer_type",
+      TileLayerType.fromString,
+      _.repr
+    )
 }